refactor(day, order): 패키지 추가, 이동
* Day, DayOfWeek: 날짜에 관한 도메인은 day 패키지
* OrderSheet, Menu, MenuType: 주분에 관련한 도메인은 order 패키지로 이동
