+++ src/main/java/christmas/Application.java
@@ -1,5 +1,6 @@
 package christmas;
 
+import christmas.controller.PromotionController;
 import christmas.view.InputView;
 import christmas.view.OutputView;
