+++ src/main/java/christmas/domain/day/Day.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.day;
 
 import java.util.List;
