+++ src/main/java/christmas/domain/day/DayOfWeek.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.day;
 
 public enum DayOfWeek {
