+++ src/main/java/christmas/domain/event/presentation/PresentationEventManager.java
@@ -1,11 +1,10 @@
 package christmas.domain.event.presentation;
 
-import christmas.domain.Menu;
-import christmas.domain.OrderSheet;
+import christmas.domain.order.Menu;
+import christmas.domain.order.OrderSheet;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.Function;
 
 public enum PresentationEventManager {
     CHAMPAGNE {
