+++ src/main/java/christmas/domain/order/MenuType.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 public enum MenuType {
     MAIN, DESERT, APPETIZER, DRINK
