+++ src/test/java/christmas/DayTest.java
@@ -1,6 +1,6 @@
 package christmas;
 
-import christmas.domain.Day;
+import christmas.domain.day.Day;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
