+++ src/test/java/christmas/event/discount/DiscountResultTest.java
@@ -1,8 +1,7 @@
 package christmas.event.discount;
 
-import christmas.domain.Day;
-import christmas.domain.Menu;
-import christmas.domain.OrderSheet;
+import christmas.domain.day.Day;
+import christmas.domain.order.OrderSheet;
 import christmas.domain.event.discount.*;
 import org.junit.jupiter.api.Test;
