+++ src/test/java/christmas/event/discount/WeekDayDiscountEventTest.java
@@ -1,7 +1,7 @@
 package christmas.event.discount;
 
-import christmas.domain.Day;
-import christmas.domain.OrderSheet;
+import christmas.domain.day.Day;
+import christmas.domain.order.OrderSheet;
 import christmas.domain.event.discount.WeekDayDiscountEvent;
 import org.junit.jupiter.api.Test;
